fix(mobile): repair import paths broken by the file reorganization
90882d0 ("mobile: reorganized files", 2024-12-08) moved storage.ts and
screens.ts into utils/ and styles.variables.ts into styles/, but left 23
imports pointing at the old locations. Metro resolves nothing through
aliases here — the config is expo's default — so the app has not been able
to bundle since that commit.

Three shapes of breakage:
- ./storage and ./screens now live under utils/
- ../styles.variables now lives at ../styles/styles.variables
- screens/*/ files referenced ../components and ../utils, one level short
  of ../../

Also fixes ../Storage, which only resolved on case-insensitive filesystems.

tsc now reports no module resolution errors; the 7 that remain are all
nested objects inside StyleSheet.create and are unrelated.
